+++ J_RachioDevice1.js
@@ -15,7 +15,7 @@ var RachioDevice1 = (function (api) {
 	// unique identifier for this plugin...
 	var uuid = "ccdca3bc-6fca-11e7-b3ef-74d4351650de";
 
-	var pluginVersion = '1.7develop-20300';
+	var pluginVersion = '1.7develop-21165.1600';
 
 	var serviceId = "urn:toggledbits-com:serviceId:RachioDevice1";